+++ CHANGELOG.md
@@ -18,7 +18,6 @@ Released: TBA.
 [Diff](https://github.com/kvz/lanyon/compare/v0.0.116...master).
 
 - [ ] Add a 'real' command line parser for `lib/cli.js` like minimist
-- [ ] Remove `shelljs` completely
 - [ ] Fix bug where failed deploy is not fatal: https://travis-ci.org/kvz/invig/builds/202931498#L627
 - [ ] Make it so that you can only build e.g. a homepage via `LANYON_EXCLUDE=* LANYON_INCLUDE=home.html,_layouts/default.html`. However, we first need this Jekyll issue resolved: https://github.com/jekyll/jekyll/issues/4791#issuecomment-289021488
 - [ ] Bundle node + modules in docker container also (and see if we can use them, using `open` for browsersync)
@@ -31,6 +30,8 @@ Released: TBA.
 Released: 2018-04-12.
 [Diff](https://github.com/kvz/lanyon/compare/v0.0.115...v0.0.116).
 
+- [x] Remove `shelljs` completely
+- [x] Deprecate magic themeDir handling for now
 - [x] Use pkill instead of killall so it works on both Linux & OSX (thanks @Acconut)
 - [x] Restart lanyon container if it crashes
 - [x] Cleaner kills for cleanup